requirements.txt
ruamel.yaml
pandas
scipy
torch
torchvision
numpy
easydict
matplotlib